backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-socketio==5.5.1
eventlet==0.39.1
bcrypt==4.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
